requirements.txt
crewai
tavily-python
google-genai
python-dotenv
requests
beautifulsoup4
lxml
litellm
